refactor(velite): move posts definition to `content/collections/posts.ts`
