+++ content/plugin/rehype-autolink-headings.ts
@@ -1,9 +1,8 @@
-import pluggin, { Options } from 'rehype-autolink-headings'
+import plugin, { Options } from 'rehype-autolink-headings'
 
 const options: Partial<Options> = {
   behavior: 'wrap'
 }
 
-const rehypeAutolinkHeadings = { pluggin, options }
-
+const rehypeAutolinkHeadings = [plugin, options]
 export default rehypeAutolinkHeadings
