+++ contentlayer.config.ts
@@ -2,7 +2,6 @@
 
 import { makeSource } from 'contentlayer/source-files'
 
-import { Post } from './content/definitions/Post'
 import { Project } from './content/definitions/Project'
 import { Til } from './content/definitions/Til'
 import { AboutMe } from './content/definitions/AboutMe'
@@ -10,7 +9,7 @@ import { remarkPlugins, rehypePlugins } from './content/plugin'
 
 export default makeSource({
   contentDirPath: 'content',
-  documentTypes: [Post, Project, Til, AboutMe],
+  documentTypes: [Project, Til, AboutMe],
   disableImportAliasWarning: true,
   mdx: {
     remarkPlugins,
